+++ R/utils.R
@@ -723,6 +723,13 @@ verify_key_type <- function(p) {
   for (i in seq_along(keys)) {
     k <- keys[[i]]
     if (is.null(k)) next
+    if ("select" %in% p$x$layout$clickmode && "plotly_click" %in% p$x$highlight$on) {
+      warning(
+        "`layout.clickmode` = 'select' is not designed to work well with ",
+        "the R package's linking framework (i.e. crosstalk support).",
+        call. = FALSE
+      )
+    }
     # does it *ever* make sense to have a missing key value?
     uk <- uniq(k)
     if (length(uk) == 1) {